requirements.txt
asgiref==3.9.1
Django==4.2.11
django-ckeditor==6.7.0
django-js-asset==3.1.2
gunicorn==23.0.0
packaging==25.0
Pillow==10.1.0
sqlparse==0.5.3
